it's somewhat decent looking now, made shops the front page and yeah that's about it..
