+++ models/shop.js
@@ -2,7 +2,8 @@ var mongoose = require('mongoose');
 var Schema = mongoose.Schema;
 
 const shopSchema = new Schema({
-  name: String
+  name: String,
+  description: String
 }, {timestamps: true});
 
 module.exports = mongoose.model('Shop', shopSchema);
