+++ routes/index.js
@@ -1,43 +1,8 @@
 var express = require('express');
-const passport = require('passport');
 const { ensureAuthenticated } = require('../strategies/auth');
-const handlebars = require('hbs');
 var router = express.Router();
 var Shops = require('../models/shop');
 var Items = require('../models/items');
-
-/* GET home page test - make sure you only have one of the '/' routes un-commented. */
-
-// router.get('/', function (req, res) {
-//   res.send('Hello World!')
-// });
-
-handlebars.registerHelper("iteminbasket", function(itemid, options){
-  if(options.data.root.user){
-    var result = options.data.root.user.basket.find( x => x.item == itemid);
-    if(result){
-      return result.quantity;
-    }else{
-      return null;
-    }
-  }else{
-    return null;
-  }
-});
-
-handlebars.registerHelper("convertPrice", function(price){
-  var sPrice = price.toString();
-  var length = sPrice.length;
-
-  // console.log(sPrice + ":" + length);
-  if (length <= 2){
-    return "£0." + sPrice;
-  }else{
-    sPrice = [sPrice.slice(0, length - 2), ".", sPrice.slice(length - 2)].join('');
-    return "£" + sPrice;
-    // console.log(sPrice);
-  }
-});
 
 router.get('/', async function(req, res, next) {
   var shops = await Shops.find(); //let's get all the shops
