+++ routes/shops.js
@@ -18,6 +18,7 @@ router.post('/add', (req, res) =>{
 });
 
 router.get('/:id', async (req, res) =>{
+    console.log('got something');
     var id = req.params.id;
     var items = await Items.find({shop: id});
     Shops.findOne({_id: id}).then(data => {
