--- views/partials/nav.hbs
@@ -1,16 +1,0 @@
-<nav>
-    <h1><a href="/">Deliver2</a></h1>
-    {{#if user}}
-    <p>{{user.email}}</p>
-    <div class="container-row center">
-        <a href="/users/logout" class="button-white">logout</a>
-        <a href="/users/basket"><span class="material-icons" style="font-size: 2rem;">shopping_cart</span></a>
-    </div>
-    
-    {{else}}
-    <ul>
-        <li class="button-white"><a href="/users/login" style="color: black;">Login</a></li>
-        <li class="button-white"><a href="/users/register" style="color: black;">Register</a></li>
-    </ul>
-    {{/if}}
-</nav>
